Simplify bug/exception finding in parallel code
This commit is triggered by a bug (that is also fixed by it).
There, we forget the tuple-brackets when putting a job onto the
job queue. Then the job that encountered the exception finished.
But since it did not put its resources back onto the required
queues, other jobs were deadlocked.
Consequently, the pool was never able to collect the workers,
and we were never alerted about errors in the parallel code.

In the redesigned structure, an exception in any of the workers
will gracefully shut down all workers. Now we can see the error
exception message, fix it and restart the monitoring loop. Note
that this change should only be relevant during development.
The recently introduced tests on synthetic data should catch
those kinds of errors.